\name{neworder-package}
\alias{neworder-package}
\alias{neworder}
\docType{package}
\title{
 A C++ microsimulation package for R, python, and (probably) node.js
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Andrew P Smith <a.p.smith@leeds.ac.uk>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
